+++ src/main/java/nextstep/ladder/LadderGameApplication.java
@@ -1,5 +1,7 @@
 package nextstep.ladder;
 
+import nextstep.ladder.domain.*;
+
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
